Milestone 1
This code DOES NOT work on my machine. Odd issues with function calls being invalid.
Continuing to explore, but we can use this as a baseline for 408 development.
To be clear, this is NOT my code and I DO NOT take credit for writing it... the pictures are mine though! :)