Add files via upload
Very basic example of sending a char (currently 'L') from Python to arduino. It seems to enter the if(inChar == 'L') statement in the arduino code (or the Python code wouldn't close), but it doesn't always print to the Serial Monitor There is a problem when switching the serial destination (from Python program switching to the Serial Monitor).